Ensure the test process prints to stdout before exiting
